+++ scripts/appendTooltipTo.js
@@ -1,14 +1,14 @@
 module.exports = function($el, errors, tooltipId) {
   let $body = $el.closest('body');
-  let existingTooltipId = $el.attr('html-linter-tooltip-id');
+  let existingTooltipId = $el.attr('layout-linter-tooltip-id');
 
   let $tooltip;
 
   if (existingTooltipId) {
     $tooltip = $body.find(`#${existingTooltipId}`);
   } else {
-    $el.attr('html-linter-tooltip-id', tooltipId);
-    $body.append(`<div id="${tooltipId}" class="html-linter-tooltip"><ul></ul></div>`);
+    $el.attr('layout-linter-tooltip-id', tooltipId);
+    $body.append(`<div id="${tooltipId}" class="layout-linter-tooltip"><ul></ul></div>`);
     $tooltip = $body.find(`#${tooltipId}`);
   }