+++ tests/modules/fragment-option-test.js
@@ -3,19 +3,19 @@
 const assert = require('assert');
 
 const createScenario = require('../helpers/createScenario');
-const lintHtml = require('../../index');
+const lintLayout = require('../../index');
 
 describe('testing fragment option', function() {
   it('will return whole linted HTML document when passed whole HTML document', function() {
     createScenario({
-      relativePathToRulesFile: '/.htmllintrc',
+      relativePathToRulesFile: '/.layoutrc',
       rules: [{
         selector: '.test',
         is: 'p'
       }]
     });
 
-    const result = lintHtml({
+    const result = lintLayout({
       source: `
         <html>
           <head>
@@ -54,14 +54,14 @@ describe('testing fragment option', function() {
 
   it('will return whole linted HTML document when passed an HTML fragment', function() {
     createScenario({
-      relativePathToRulesFile: '/.htmllintrc',
+      relativePathToRulesFile: '/.layoutrc',
       rules: [{
         selector: '.test',
         is: 'p'
       }]
     });
 
-    const result = lintHtml({
+    const result = lintLayout({
       source: `
         <div id="some-original-element"></div>
         <div class="test"></div>
@@ -92,14 +92,14 @@ describe('testing fragment option', function() {
 
   it('will return the linted fragment when passed a fragment, if `fragment` option is set to `true`', function() {
     createScenario({
-      relativePathToRulesFile: '/.htmllintrc',
+      relativePathToRulesFile: '/.layoutrc',
       rules: [{
         selector: '.test',
         is: 'p'
       }]
     });
 
-    const result = lintHtml({
+    const result = lintLayout({
       source: `
         <div id="some-original-element"></div>
         <div class="test"></div>
@@ -123,14 +123,14 @@ describe('testing fragment option', function() {
 
   it('will return whole linted HTML document, even if `fragment` option is set to `true`, if the "fragment" starts with <!doctype', function() {
     createScenario({
-      relativePathToRulesFile: '/.htmllintrc',
+      relativePathToRulesFile: '/.layoutrc',
       rules: [{
         selector: '.test',
         is: 'p'
       }]
     });
 
-    const result = lintHtml({
+    const result = lintLayout({
       source: `
         <!doctype html>
         <html>
@@ -165,14 +165,14 @@ describe('testing fragment option', function() {
 
   it('will return whole linted HTML document, even if `fragment` option is set to `true`, if the "fragment" starts with <html', function() {
     createScenario({
-      relativePathToRulesFile: '/.htmllintrc',
+      relativePathToRulesFile: '/.layoutrc',
       rules: [{
         selector: '.test',
         is: 'p'
       }]
     });
 
-    const result = lintHtml({
+    const result = lintLayout({
       source: `
         <html>
           <div id="some-original-element"></div>
@@ -206,14 +206,14 @@ describe('testing fragment option', function() {
 
   it('will return whole linted HTML document, even if `fragment` option is set to `true`, if the "fragment" starts with <head', function() {
     createScenario({
-      relativePathToRulesFile: '/.htmllintrc',
+      relativePathToRulesFile: '/.layoutrc',
       rules: [{
         selector: '.test',
         is: 'p'
       }]
     });
 
-    const result = lintHtml({
+    const result = lintLayout({
       source: `
         <head></head>
         <div id="some-original-element"></div>
@@ -246,14 +246,14 @@ describe('testing fragment option', function() {
 
   it('will return whole linted HTML document, even if `fragment` option is set to `true`, if the "fragment" starts with <body', function() {
     createScenario({
-      relativePathToRulesFile: '/.htmllintrc',
+      relativePathToRulesFile: '/.layoutrc',
       rules: [{
         selector: '.test',
         is: 'p'
       }]
     });
 
-    const result = lintHtml({
+    const result = lintLayout({
       source: `
         <body>
           <div id="some-original-element"></div>